pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-judge"
version = "0.1.0"
description = "LLM-as-a-Judge evaluation framework for text-to-video models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Daria"}
]
keywords = ["video", "generation", "llm", "evaluation", "benchmark", "judge"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "fal-client>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.4.1",
]

[project.scripts]
video-judge = "video_judge.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/video-judge"
Repository = "https://github.com/yourusername/video-judge"
Issues = "https://github.com/yourusername/video-judge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["video_judge*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
